requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.7
colorama==0.4.6
fastapi==0.104.0
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
idna==3.4
loguru==0.7.2
Mako==1.2.4
MarkupSafe==2.1.3
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
win32-setctime==1.1.0
